+++ src/meson.build
@@ -33,6 +33,7 @@ headers = [
 privates = [
   'fw_iso_ctx_private.h',
   'fw_iso_resource_private.h',
+  'fw_iso_resource_private.c',
 ]
 
 inc_dir = meson.project_name()
